Use u32

+++ src/animation/Skeleton.h
@@ -48,7 +48,7 @@ struct BoneTransform {
 };
 
 struct Bone {
-	std::vector<size_t> idxvertices;
+	std::vector<u32>  idxvertices; // TODO use u16 here ?
 	long              father;
 	
 	BoneTransform     anim;
